[stm32l475] improve the structrue of sconscript
