+++ bsp/stm32/stm32l475-atk-pandora/applications/SConscript
@@ -7,20 +7,16 @@ src = Split('''
 main.c
 ''')
 
-if GetDepend(['BSP_USING_LCD_SAMPLE']):
-    src += ['lcd_sample.c']
-
 if GetDepend(['PKG_USING_NRF24L01']):
-    src += ['nrf24l01_init.c']
+    src += Glob('nrf24l01_init.c')
 
 CPPPATH = [cwd]
 
 group = DefineGroup('Applications', src, depend = [''], CPPPATH = CPPPATH)
 
 list = os.listdir(cwd)
-for d in list:
-    path = os.path.join(cwd, d)
-    if os.path.isfile(os.path.join(path, 'SConscript')):
-        group = group + SConscript(os.path.join(d, 'SConscript'))
+for item in list:
+    if os.path.isfile(os.path.join(cwd, item, 'SConscript')):
+        group = group + SConscript(os.path.join(item, 'SConscript'))
 
 Return('group')
