+++ bsp/stm32/stm32l475-atk-pandora/board/Kconfig
@@ -41,6 +41,13 @@ menu "Onboard Peripheral Drivers"
         depends on BSP_USING_SPI_LCD && !BSP_USING_LVGL
         default n
 
+    config BSP_USING_LCD_QRCODE
+        bool "Enable LCD to show QRCode"
+        depends on BSP_USING_SPI_LCD && !BSP_USING_LVGL
+        select BSP_USING_LCD_SAMPLE
+        select PKG_USING_QRCODE
+        default n
+
     config BSP_USING_LVGL
         bool "Enable LVGL for LCD"
         select PKG_USING_LVGL
