+++ bsp/stm32/stm32l475-atk-pandora/board/SConscript
@@ -12,15 +12,12 @@ board.c
 CubeMX_Config/Src/stm32l4xx_hal_msp.c
 ''')
 
-if GetDepend('BSP_USING_KEY'):
-    src = src + ['ports/drv_key.c']
+if GetDepend(['BSP_USING_KEY']):
+    src += Glob('ports/drv_key.c')
 
 if GetDepend(['BSP_USING_QSPI_FLASH']):
     src += Glob('ports/drv_qspi_flash.c')
     
-if GetDepend('BSP_USING_SPI_LCD'):
-    src = src + ['ports/drv_lcd.c']
-
 if GetDepend(['BSP_USING_SDCARD']):
     src += Glob('ports/drv_sdcard.c')
 
@@ -39,7 +36,6 @@ if GetDepend(['BSP_USING_STM32_SDIO']):
 
 path =  [cwd]
 path += [cwd + '/CubeMX_Config/Inc']
-path += [cwd + '/ports']
 
 if GetDepend(['BSP_USING_AUDIO']):
     path += [cwd + '/ports/audio']
@@ -55,11 +51,5 @@ elif rtconfig.CROSS_TOOL == 'iar':
 
 CPPDEFINES = ['STM32L475xx'] 
 group = DefineGroup('Drivers', src, depend = [''], CPPPATH = path, CPPDEFINES = CPPDEFINES)
-
-list = os.listdir(cwd)
-for d in list:
-    path = os.path.join(cwd, d)
-    if os.path.isfile(os.path.join(path, 'SConscript')):
-        group = group + SConscript(os.path.join(d, 'SConscript'))
 
 Return('group')
