+++ bsp/stm32/stm32l475-atk-pandora/board/ports/lcd/SConscript
@@ -1,0 +1,16 @@
+from building import *
+import os
+
+cwd = GetCurrentDir()
+src = Glob('*.c')
+CPPPATH = [cwd]
+
+if GetDepend(['BSP_USING_LCD_QRCODE']):
+    src = src + Glob('lcd_qrcode.c')
+
+if GetDepend(['BSP_USING_LCD_SAMPLE']):
+    src = src + Glob('demo/lcd_sample.c')
+
+group = DefineGroup('Drivers', src, depend = ['BSP_USING_SPI_LCD'], CPPPATH = CPPPATH)
+
+Return('group')
